backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
bleach==3.3.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.35.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
distlib==0.3.7
distro==1.8.0
dj-database-url==2.1.0
Django==4.1
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-countries==7.5.1
django-country==0.1.0
django-extensions==3.2.3
django-filter==23.3
django-js-asset==2.1.0
django-qr-code==3.1.1
django-quill-editor==0.1.40
django-summernote==0.8.20.0
django-taggit==5.0.0
django-tinymce==3.6.1
djangorestframework==3.14.0
easy-thumbnails==2.8.5
environ==1.0
filelock==3.12.2
-e git+https://github.com/yeahwhat-mc/goslate@d6d511a9c001fe0b5c1cf947fd650770e9794f1d#egg=goslate
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
humanize==4.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iso8601==0.1.12
joblib==1.3.2
libretranslatepy==2.1.1
linecache2==1.0.0
lxml==4.9.3
mailjet-rest==1.3.4
msgpack==1.0.7
mysql==0.0.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
nltk==3.8.1
openai==1.3.8
packaging==23.2
pbr==5.11.1
Pillow==10.0.1
pipenv==2023.7.23
platformdirs==3.10.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDictionary==2.0.1
pyOpenSSL==23.3.0
pypng==0.20220715.0
pytermii==0.0.3
pytz==2019.3
qrcode==7.4.2
redis==5.0.1
regex==2023.10.3
requests==2.31.0
segno==1.5.3
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stevedore==5.1.0
termii==0.1.1
termii-sdk==1.0
tqdm==4.66.1
traceback2==1.4.0
translate==3.6.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.16
virtualenv==20.24.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
whitenoise==6.6.0
zope.interface==6.1
